+++ src/animations/player.ts
@@ -85,3 +85,19 @@ export function addAnimationController(
 ) {
   return new AnimationController(host, target);
 }
+
+type ViewTransitionResult = {
+  transition?: ViewTransition;
+};
+
+export function startViewTransition(
+  callback?: ViewTransitionUpdateCallback
+): ViewTransitionResult {
+  /* c8 ignore next 4 */
+  if (getPrefersReducedMotion() || !document.startViewTransition) {
+    callback?.();
+    return {};
+  }
+
+  return { transition: document.startViewTransition(callback) };
+}