+++ src/index.ts
@@ -56,6 +56,8 @@ export { default as IgcSliderLabelComponent } from './components/slider/slider-l
 export { default as IgcTabsComponent } from './components/tabs/tabs.js';
 export { default as IgcTabComponent } from './components/tabs/tab.js';
 export { default as IgcTabPanelComponent } from './components/tabs/tab-panel.js';
+export { default as IgcTileComponent } from './components/tile-manager/tile.js';
+export { default as IgcTileManagerComponent } from './components/tile-manager/tile-manager.js';
 export { default as IgcToastComponent } from './components/toast/toast.js';
 export { default as IgcToggleButtonComponent } from './components/button-group/toggle-button.js';
 export { default as IgcSwitchComponent } from './components/checkbox/switch.js';
@@ -107,6 +109,7 @@ export type { IgcSnackbarComponentEventMap } from './components/snackbar/snackba
 export type { IgcStepperComponentEventMap } from './components/stepper/stepper.common.js';
 export type { IgcTabsComponentEventMap } from './components/tabs/tabs.js';
 export type { IgcTextareaComponentEventMap } from './components/textarea/textarea.js';
+export type { IgcTileComponentEventMap } from './components/tile-manager/tile.js';
 export type { IgcTreeComponentEventMap } from './components/tree/tree.common.js';
 
 // Public types
@@ -114,7 +117,10 @@ export type {
   RangeTextSelectMode,
   SelectionRangeDirection,
   StyleVariant,
+  TileManagerDragMode,
+  TileManagerResizeMode,
 } from './components/types.js';
+export type { IgcTileChangeState } from './components/tile-manager/tile.js';
 export type {
   DateRangeDescriptor,
   WeekDays,